Update lambda deploy script
I personally have a feeling that the way pnpm is trying to install might be breaking things, but I could be very wrong. Instead of relying on the lockfile, I will try to force a delete of node_modules and the lockfile, and install new. This may or may not work, although I believe it won't, but it is better than installing missing modules one by one when I don't know how many there are.