Change from pnpm to npm
After a lot of research, it appears that the main reason for the 502 bad gateway errors were due to the way pnpm deals with node_modules. By converting to npm, this should increase the compatibility with AWS Lambda.

For the sake of consistency, I have also converted the frontend to use npm as well, to make working with a monorepo simpler, rather than going for a hybrid approach.